fix: 3d to 3i

+++ main/src/main/java/org.echoesfrombeyond.jam/Plugin.java
@@ -42,7 +42,7 @@ import org.jspecify.annotations.Nullable;
 public class Plugin extends JavaPlugin {
   public static final Vector3d VEC = new Vector3d(1127, 143, -2685);
 
-  public static final Vector3d RADIO_TOWER_LOC = new Vector3d(1127, 143, -2685);
+  public static final Vector3i RADIO_TOWER_LOC = new Vector3i(1127, 143, -2685);
 
   private static @Nullable ResourceType<ChunkStore, JamSave> JAM_TYPE;
   private static @Nullable ComponentType<EntityStore, PlacePreviewComponent> PLACE_TYPE;
